+++ pkg/ring/token_generator.go
@@ -136,7 +136,11 @@ func (g *MinimizeSpreadTokenGenerator) GenerateTokens(ring *Desc, id, zone strin
 	for i := 1; i <= len(zonalTokens); i++ {
 		index := i % len(zonalTokens)
 		if tokenInstanceId, ok := usedTokens[zonalTokens[index]]; ok && tokenInstanceId != id {
-			instanceDistance := tokensPerInstanceWithDistance[tokenInstanceId]
+			instanceDistance, ok := tokensPerInstanceWithDistance[tokenInstanceId]
+			if !ok {
+				continue // Same token is shared to an ingester in different zone, skip
+			}
+
 			instanceDistance.tokens = append(instanceDistance.tokens, &tokenDistanceEntry{
 				token:    zonalTokens[index],
 				prev:     zonalTokens[i-1],