+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * [BUGFIX] Ingester: Avoid resharding for query when restart readonly ingesters. #6642
 * [BUGFIX] Query Frontend: Fix query frontend per `user` metrics clean up. #6698
 * [BUGFIX] Add `__markers__` tenant ID validation. #6761
+* [BUGFIX] Ring: Fix nil pointer exception when token is shared. #6768
 
 ## 1.19.0 2025-02-27